Cleanup box selecting
Noticed a bug in Studio: if you shift+dblclick to zoom out,
then click again, Draw thinks you're box selecting. Had to
make all clicks cancel box select. To do that, had to allow
multiple handlers to be called on the same event -- do not
see a reason that should not be allowed.