Enable touch event propagation properly
This is to fix #617
It allows the event to propagate through if there are no mapbox targets nearby.
This sort of breaks the one tap move drawing functionality, but still allows for the double tap to move the drawing.
It also enables popups and clusters to work again.

I believe this is the best temporary solution until another is found.  The one provided by @Plantain here:  https://github.com/mapbox/mapbox-gl-draw/pull/830
basically broke drawing completely on some browsers.  I have tested extensively and am employing the patch to all my apps.