Fix selection event in direct_select w/ coordPath
The selectionchange event fired when moving into direct_select mode with
a coordPath will now correctly include the point. Before, the event was
fired with the correct selected feature, but lacking in the point.

This bug was caused because of a misordered selection call: First the
coordinates were selected, and only then the feature. During the feature
selection, the store's refreshSelectedCoordinates function was called
*before* the feature was actually selected, meaning the previously
coordinates were effectively removed.