Include shims for the fs and path dependencies
Including fs and path in externals is not sufficient when this is used
in a browser. That will just make rollup leave the requires calls, and
when you try to include this library in a webpack project webpack will
complain that fs is missing.

By providing shims for fs and path, those will be used instead of
require calls, so that fixes the issue. Since no code paths in this
library triggers the code that requires fs or path it's safe to just
shim them out with an empty object.

Fixes #970