fix: trigger featureChanged on feature props change
https://github.com/mapbox/mapbox-gl-draw/pull/1196 intended to add a call to
`featureChanged` any time a feature was `add()`ed with modified
properties, however it has a bug in that the check came after assignment.

A simpler fix would be to move the assignment inside the `if()` body,
but that could break other code that depended on referential equality
after calling `add()`.